+++ GPS/GpsData.h
@@ -18,6 +18,6 @@ typedef union {
     uint32_t timeStamp_ms;
   };
   
-  uint16_t values[GPS_DATA_SIZE_BYTE];
+  uint8_t values[GPS_DATA_SIZE_BYTE];
 }
 GpsData;
